+++ sdk/src/api/embeddedWallet.ts
@@ -19,6 +19,7 @@ import {
   OpenfortEvents,
   ChainTypeEnum,
   AccountTypeEnum,
+  ListAccountsParams,
 } from '../types/types';
 import { TypedDataPayload } from '../wallets/evm/types';
 import { IframeManager } from '../wallets/iframeManager';
@@ -323,7 +324,8 @@ export class EmbeddedWalletApi {
 
     const signer = await this.ensureSigner();
     const { hashMessage = true, arrayifyMessage = false } = options || {};
-    return await signer.sign(message, arrayifyMessage, hashMessage);
+    const account = await Account.fromStorage(this.storage);
+    return await signer.sign(message, arrayifyMessage, hashMessage, account?.chainType);
   }
 
   async signTypedData(
@@ -406,7 +408,11 @@ export class EmbeddedWalletApi {
     };
   }
 
-  async list(): Promise<EmbeddedAccount[]> {
+  async list(requestParams?: ListAccountsParams): Promise<EmbeddedAccount[]> {
+    const params = {
+      accountType: AccountTypeEnum.SMART_ACCOUNT,
+      ...requestParams,
+    };
     const configuration = SDKConfiguration.fromStorage();
     if (!configuration) {
       throw new OpenfortError('Configuration not found', OpenfortErrorType.INVALID_CONFIGURATION);
@@ -418,9 +424,7 @@ export class EmbeddedWalletApi {
     }
     return withOpenfortError<EmbeddedAccount[]>(async () => {
       const response = await this.backendApiClients.accountsApi.getAccountsV2(
-        {
-          accountType: AccountTypeEnum.SMART_ACCOUNT,
-        },
+        params,
         {
           headers: {
             authorization: `Bearer ${configuration.baseConfiguration.publishableKey}`,