+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.12] - 2025-08-14
+### Fix
+- Solana sample
+
+### Feat
+- sign method uses chain type from the local storage
+
 ## [0.9.8] - 2025-08-14
 ### Feat
 - Improve logout handling